sort

+++ BlogClientApp/src/app/articles/aricle-details/aricle-details.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Post, ArticleService } from 'src/app/shared/article.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { AdditionalDataService, AdditionalInfo } from 'src/app/shared/additional-data.service';
+import { Category, CategoryService } from 'src/app/shared/category.service';
+import { Tag, TagService } from 'src/app/shared/tag.service';
 
 @Component({
   selector: 'app-aricle-details',
@@ -13,10 +14,30 @@ export class AricleDetailsComponent implements OnInit {
   private id: number;
   articleItem:Post[] ;
   isAdmin = false;
+  isEdit = false;
 
-  //additionalData:AdditionalInfo[];
+  notChangedCategory = true;
+  notChangedTags = true;
+
+  //Для редактирования статьи
+  editedArticle: Post = {
+    articleId: undefined,
+    categoryId: undefined,
+    categoryName:"",
+    description:"",
+    heroImage:"",
+    name:"",
+    publicationDate: undefined,
+    shortDescription:"",
+    tags: null
+  };
   
-  constructor(private route: ActivatedRoute, private service: ArticleService, private router: Router, private additionalService: AdditionalDataService ) { }
+  categories: Category[];
+  tags: Tag[];
+
+
+
+  constructor(private route: ActivatedRoute, private service: ArticleService, private router: Router, private tagService: TagService, private categoryService: CategoryService) { }
 
   ngOnInit() {
     if(localStorage.getItem('token') != null)
@@ -29,8 +50,6 @@ export class AricleDetailsComponent implements OnInit {
       this.getArticleById();
       //this.getAdditionalData();
     });
-
-    
   }
 
   getArticleById(){
@@ -64,12 +83,70 @@ export class AricleDetailsComponent implements OnInit {
   }
 
   deletePost(){
-    this.service.deletePost(this.id).subscribe(data =>{ 
-      this.router.navigateByUrl('/admin/adminarticles');
+    let del = confirm("Удалить статью?");
+
+    if(del){
+      this.service.deletePost(this.id).subscribe(data =>{ 
+        this.router.navigateByUrl('/admin/adminarticles');
+      },
+      err => {
+        console.log(err);
+      });
+    }
+  }
+
+  editArticle(){
+    this.isEdit = !this.isEdit;
+
+    if(!this.isEdit){
+      return;
+    }
+
+    window.scrollTo(0,0);
+
+    if(this.tags == null || this.tags == undefined){
+      this.getTags();
+    }
+    if(this.categories == null || this.categories == undefined){
+      this.getCategories();
+    }
+    
+    this.editedArticle.articleId = this.articleItem[0].articleId;
+    this.editedArticle.categoryId = this.articleItem[0].categoryId;
+    this.editedArticle.name = this.articleItem[0].name;
+    this.editedArticle.shortDescription = this.articleItem[0].shortDescription;
+    this.editedArticle.description = this.articleItem[0].description;
+    this.editedArticle.heroImage = this.articleItem[0].heroImage;
+  
+    console.log(this.editedArticle);
+  }
+
+  getTags(){
+    this.tagService.getTags().subscribe(data => {
+      this.tags = data;
+      this.editedArticle.tags = this.articleItem[0].tags;
+      console.log(this.tags);
     },
-    err => {
+    err =>{
       console.log(err);
-    });
+    })
+  }
+
+  getCategories(){
+    this.categoryService.getCategories().subscribe(data => {
+      this.categories = data;
+      console.log(this.categories);
+    },err => {
+      console.log(err);
+    })
+  }
+
+  changedCategory(){
+    this.notChangedCategory = false;
+  }
+
+  changedTags(){
+    this.notChangedTags = false;
   }
 
 }
